fix: split token_scan to fix norminette
