+++ parser/parser.h
@@ -67,6 +67,7 @@ bool				match_word(char *src, size_t *current);
 bool				match_identifier(char *src, size_t *current);
 bool				match_var(char *src, size_t *current);
 char				*extract_word(char *src, size_t *current);
+t_token				*token_identify(char *src, size_t *current);
 t_token				*extract_str(char *src, size_t *current, bool single);
 t_token				*extract_identifier(char *src, size_t *current);
 t_token				*extract_var(char *src, size_t *current);
